Add DiscordRole option and custom option for XP
